--- Team7561-MediaProject/README.md
@@ -1,16 +1,0 @@
-Name of project: Team7561-MediaProject
-Name of team members: Axel Castellanos, Fabian Santano, Cristobal Elizarraraz, Saul Machuca
-Class: CST205-01_SP23
-Date: 05/17/2023
-How to run program: 
-  Download code or make clone of repository.
-  Once that is done, use Powershell(windows) or Terminal(mac) to launch flask application by using the line "flask --app multimedia.py --debug run".
-  Message authors for database key and replace "os.environ.get('MYSQL_PASSWORD')" in line 37 of file "multimedia.py" with the database key.
-  User is then prompted to enter userr name or password to enter application or create a new account to enter the application in the future.
-  
-Link to GitHub repository: https://github.com/Cast02/Team7561-MediaProject
-Link to Trello board: https://trello.com/invite/b/DjCLNQQ3/ATTI2f60a3b4f15e3dccaab21b77628396e3AD4B8D4B/team-7561
-Future work:
-   Add better CSS
-   Add a save recipe feature, where users can save their favorite recipes.
-   Add a create recipe feature, where users can create their own recipes and they will be saved under their account
