--- Team7561-MediaProject/static/style.css
@@ -1,107 +1,0 @@
-
-html,body{
-    text-align:center;
-    background-color:dimgray;
-    min-height: 100%;
-    width: 100%;
-  }
-
-.page {
-    background: #e2e2e5;
-    display: flex;
-    flex-direction: column;
-    height: calc(100% - 40px);
-    position: absolute;
-    place-content: center;
-    width: calc(100% - 40px);
-  }
-  
-  @media (max-width: 767px) {
-    .page {
-      height: auto;
-      margin-bottom: 20px;
-      padding-bottom: 20px;
-    }
-  }
-  .container {
-    display: flex;
-    height: 320px;
-    margin: 0 auto;
-    width: 640px;
-  }
-  @media (max-width: 767px) {
-    .container {
-      flex-direction: column;
-      height: 630px;
-      width: 320px;
-    }
-  }
-  .left {
-    background: white;
-    height: calc(100% - 40px);
-    top: 20px;
-    position: relative;
-    width: 50%;
-  }
-  @media (max-width: 767px) {
-    .left {
-      height: 100%;
-      left: 20px;
-      width: calc(100% - 40px);
-      max-height: 270px;
-    }
-  }
-  .login {
-    font-size: 50px;
-    font-weight: 900;
-    margin: 50px 40px 40px;
-  }
-  .eula {
-    color: #999;
-    font-size: 14px;
-    line-height: 1.5;
-    margin: 40px;
-  }
-  .right {
-    background: #474A59;
-    box-shadow: 0px 0px 40px 16px rgba(0,0,0,0.22);
-    color: #F1F1F2;
-    position: relative;
-    width: 50%;
-  }
-  @media (max-width: 767px) {
-    .right {
-      flex-shrink: 0;
-      height: 100%;
-      width: 100%;
-      max-height: 350px;
-    }
-  }
-  .form {
-  
-    margin: 40px;
-    position: absolute;
-  }
-  .formBtn {
-    padding: 50px 0;
-    text-align: center;
-  }
-
-  .parent {
-    margin: 1rem;
-    padding: 2rem 2rem;
-    text-align: center;
-  }
-  .child {
-    display: inline-block;
-    padding: 1rem 1rem;
-    vertical-align: middle;
-  }
-
-  nav{
-    border-radius: 25px;
-    padding: 1rem 1rem;
-    background-image: linear-gradient(45deg,rgb(8, 15, 221), rgb(0, 174, 255));
-    width: 50%;
-    margin-left: 25%;
-  }
