--- Team7561-MediaProject/templates/categoryList.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    {{ bootstrap.load_css() }}
-    <title>List of foods</title>
-    <link rel="stylesheet" href="../static/style.css">
-
-    <style>
-        body{
-            text-align: center;
-        }
-    </style>
-
-</head>
-<body>
-    <nav> 
-        <a class="btn btn-outline-light" href="{{ url_for('logout') }}">Logout</a> |
-        <a class="btn btn-outline-light" href = "{{ url_for('homePage') }}">Home</a> |
-        <a href="{{url_for('search')}}" class="btn btn-outline-light">Return</a>
-    </nav>
-
-    <br><br>
-    {% for meals in data['meals']%}
-        <a href="{{url_for('foodInfo', id=meals['idMeal'], category=category)}}" class="btn btn-light">{{meals['strMeal']}}</a>
-        <br>
-        <a href="{{url_for('foodInfo', id=meals['idMeal'], category=category)}}"><img src="{{meals['strMealThumb']}}" width="300" height="200"></a>
-        <br><br>
-    {% endfor %}
-    {{ bootstrap.load_js() }}
-</body>
-</html>
