--- Team7561-MediaProject/templates/homePage.html
@@ -1,147 +1,0 @@
-<!--Store this code in 'login.html' file inside the 'templates' folder-->
-<html>
-
-<head>
-	<meta charset="UTF-8">
-	{{ bootstrap.load_css() }}
-	<title>Home Page</title>
-	<link rel="stylesheet" href="../static/style.css">
-</head>
-
-<style>
-	h2 {
-		color: black;
-	}
-
-	h1 {
-		color: black;
-	}
-
-	h4 {
-		color: black;
-	}
-
-	li {
-		color: black;
-	}
-
-	h5 {
-		color: black;
-	}
-
-	h3 {
-		color: black;
-	}
-</style>
-
-<body>
-
-	<nav>
-		<!-- <a class="btn btn-outline-dark" href="{ url_for('home') }">Home </a> -->
-		<a class="btn btn-outline-light" href="{{ url_for('logout') }}">Logout</a>
-		<a class="btn btn-outline-light" href="{{ url_for('newPage') }}">Info Page</a>
-	</nav>
-	<br>
-	<h1>Welcome!</h1>
-	<h2>Please select how you would like to search for a recipe:</h2>
-	<br><br>
-	<nav>
-		<a class="btn btn-outline-light" href="{{ url_for('search') }}">Search By Food Category</a>
-		<div class="search-container"><form method="POST" action="/home_Page">
-			{{form.csrf_token}}
-			{{form.user_input.label}}
-			{{form.user_input(size=32)}}
-			  <input type="submit" value="Search"/>
-			</form>
-		  </div>
-	</nav>
-	<br><br><br><br>
-
-	<h2>Random Recipe:</h2>
-	<h3><strong>{{data['meals'][0]['strMeal']}}</strong></h3>
-	<img src="{{data['meals'][0]['strMealThumb']}}" width="300" height="300">
-
-	<br>
-	<h2><strong>Food Category:</strong> {{data['meals'][0]['strCategory']}}</h2>
-
-	<div class="parent">
-		<div class="child" style="float: left; width: 50%;">
-			<h4><strong>Instructions: </strong><br>{{data['meals'][0]['strInstructions']}}</h4>
-		</div>
-		<div class="child" style="float: right; width: 50%;">
-			<h4><strong>Ingredients:</strong></h4>
-			<h5>
-				{% if data['meals'][0]['strIngredient1'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient1'] }}</li>
-				{% endif %}
-
-				{% if data['meals'][0]['strIngredient2'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient2'] }}</li>
-				{% endif %}
-
-				{% if data['meals'][0]['strIngredient3'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient3'] }}</li>
-				{% endif %}
-
-				{% if data['meals'][0]['strIngredient4'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient4'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient5'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient5'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient6'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient6'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient7'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient7'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient8'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient8'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient9'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient9'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient10'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient10'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient11'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient11'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient12'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient12'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient13'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient13'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient14'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient14'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient15'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient15'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient16'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient16'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient17'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient17'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient18'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient18'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient19'] != '' %}
-				<li>{{ data['meals'][0]['strIngredient19'] }}</li>
-				{% endif %}
-				{% if data['meals'][0]['strIngredient20'] != '' %}
-				<li> {{ data['meals'][0]['strIngredient20'] }}</li>
-				{% endif %}
-			</h5>
-		</div>
-	</div>
-	<br>
-
-
-
-
-</body>
-
-</html>
