--- Team7561-MediaProject/templates/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    {{ bootstrap.load_css() }}
-    <title>Starting Page</title>
-</head>
-<body>
-
-
-    <h1>Welcome To Our Application</h1>
-    <h2>Is this your first time using our application?</h2>
-    <a href="{{ url_for('newUser') }}"><button>Yes</button></a>
-    <a href="{{ url_for('returing')}}"><button>No</button></a>
-</body>
-</html>
