--- Team7561-MediaProject/templates/introPage.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  {{ bootstrap.load_css() }}
-  <title>New User</title>
-  <link rel="stylesheet" href="/static/style.css">
-
-
-  <style>
-    h2 {
-      color: black;
-    }
-
-    h1 {
-      color: black;
-    }
-
-    h4 {
-      color: black;
-    }
-
-    li {
-      color: black;
-    }
-
-    h5 {
-      color: black;
-    }
-
-    h3 {
-      color: black;
-    }
-  </style>
-
-
-</head>
-
-<body>
-  <nav>
-    <!-- <a class="btn btn-outline-dark" href="{ url_for('home') }">Home </a> -->
-    <a class="btn btn-outline-light" href="{{ url_for('logout') }}">Logout</a>
-    <a class="btn btn-outline-light" href="{{ url_for('homePage') }}">Home Page</a>
-  </nav>
-  <br>
-  <h1>Welcome to our site!</h1>
-  <h2>In our application you will be able to search up recipes by different food categories.<br>For example:<br></h2>
-  <img src="static/images/categoryExample.png" width="50%" height="50%">
-  <br>
-  <h2>Once you click on a food category a list of foods will be displayed. <br>Click on one that intrests you!</h2><br>
-  <img src="static/images/listOfFoods.png" width="50%" height="50%">
-  <br><br>
-  <h2>After clicking on a food. Information about that specfic food will be displayed.</h2><br>
-  <img src="static/images/foodInfo.png" width="50%" height="50%">
-  <br><br>
-</body>
-
-</html>
