--- Team7561-MediaProject/templates/login.html
@@ -1,39 +1,0 @@
-<!--Store this code in 'login.html' file inside the 'templates' folder-->
-<html>
-
-<head>
-	<meta charset="UTF-8">
-	<title> Login </title>
-	<link rel="stylesheet" href="../static/style.css">
-</head>
-
-<body>
-	<div class="page">
-		<div class="container">
-			<!-- <h1>Login Page</h1>
-			</div></br></br></br></br> -->
-			<div class="left">
-				<div class="login">Login!</div>
-				<div class="eula">Welcome! Please log in.</div>
-			</div>
-			<div class="right">
-				<div class="form">
-					<form action="{{ url_for('returing')}}" method="post" autocomplete="off">
-						<div class="msg">{{ msg }}</div>
-						<input type="text" name="username" placeholder="Enter Your Username" class="textbox"
-							id="username" required></br></br>
-						<input type="password" name="password" placeholder="Enter Your Password" class="textbox"
-							id="password" required></br></br></br>
-						<input type="submit" class="btn" value="Login">
-					</form>
-					<a href="{{ url_for('newUser') }}"><button>Create Account</button></a>
-				</div>
-			</div>
-		</div>
-	</div>
-	</div>
-
-
-</body>
-
-</html>
