--- Team7561-MediaProject/templates/searchByCategory.html
@@ -1,102 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    {{ bootstrap.load_css() }}
-    <title>Search By Category</title>
-    <link rel="stylesheet" href="../static/style.css">
-
-</head>
-<body>
-    <nav> 
-        <a class="btn btn-outline-light" href="{{ url_for('logout') }}">Logout</a> |
-        <a class="btn btn-outline-light" href = "{{ url_for('homePage') }}">Home</a>
-    </nav>
-
-    <div class="parent">
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][0]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][0]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][0]['strCategory'])}}"><img src="{{ data['categories'][0]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][1]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][1]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][1]['strCategory'])}}"><img src="{{ data['categories'][1]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][2]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][2]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][2]['strCategory'])}}"><img src="{{ data['categories'][2]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][3]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][3]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][3]['strCategory'])}}"><img src="{{ data['categories'][3]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-    </div>
-
-    <div class="parent">
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][4]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][4]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][4]['strCategory'])}}"><img src="{{ data['categories'][4]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][5]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][5]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][5]['strCategory'])}}"><img src="{{ data['categories'][5]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][6]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][6]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][6]['strCategory'])}}"><img src="{{ data['categories'][6]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][7]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][7]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][7]['strCategory'])}}"><img src="{{ data['categories'][7]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-    </div>
-    
-    <div class="parent">
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][8]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][8]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][8]['strCategory'])}}"><img src="{{ data['categories'][8]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][9]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][9]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][9]['strCategory'])}}"><img src="{{ data['categories'][9]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][10]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][10]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][10]['strCategory'])}}"><img src="{{ data['categories'][10]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][11]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][11]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][11]['strCategory'])}}"><img src="{{ data['categories'][11]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-    </div>
-
-    <div class="parent">
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][12]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][12]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][12]['strCategory'])}}"><img src="{{ data['categories'][12]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-        <div class="child">
-            <a href="{{url_for('listCategory', category=data['categories'][13]['strCategory'])}}"><button type="button" class="btn btn-light">{{ data['categories'][13]['strCategory'] }}</button></a>
-            <br><br>
-            <a href="{{url_for('listCategory', category=data['categories'][13]['strCategory'])}}"><img src="{{ data['categories'][13]['strCategoryThumb'] }}" width="200" height="200"></a>
-        </div>
-    </div>
-
-    {{ bootstrap.load_js() }}
-</body>
-</html>
